# backend/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

import (
	"github.com/gin-gonic/gin"
)

// CreatePasteRequest defines model for CreatePasteRequest.
type CreatePasteRequest struct {
	Expiry      int     `json:"expiry"`
	Pass        *string `json:"pass,omitempty"`
	PassProtect bool    `json:"passProtect"`
	PasteText   string  `json:"pasteText"`
}

// GetPasteRequest defines model for GetPasteRequest.
type GetPasteRequest struct {
	Id   string  `json:"id"`
	Pass *string `json:"pass,omitempty"`
}

// Paste defines model for Paste.
type Paste struct {
	Id          *string `json:"id,omitempty"`
	Protected   *bool   `json:"protected,omitempty"`
	Text        *string `json:"text,omitempty"`
	TimeoutUnix *int64  `json:"timeoutUnix,omitempty"`
}

// GetPasteStatusJSONBody defines parameters for GetPasteStatus.
type GetPasteStatusJSONBody struct {
	Id string `json:"id"`
}

// GetPasteJSONRequestBody defines body for GetPaste for application/json ContentType.
type GetPasteJSONRequestBody = GetPasteRequest

// GetPasteStatusJSONRequestBody defines body for GetPasteStatus for application/json ContentType.
type GetPasteStatusJSONRequestBody GetPasteStatusJSONBody

// CreatePasteJSONRequestBody defines body for CreatePaste for application/json ContentType.
type CreatePasteJSONRequestBody = CreatePasteRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Read a paste
	// (POST /read)
	GetPaste(c *gin.Context)
	// Get last created paste ID from session
	// (GET /session)
	GetSession(c *gin.Context)
	// Get paste status (check if protected)
	// (POST /status)
	GetPasteStatus(c *gin.Context)
	// Create a new paste
	// (POST /submit)
	CreatePaste(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetPaste operation middleware
func (siw *ServerInterfaceWrapper) GetPaste(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPaste(c)
}

// GetSession operation middleware
func (siw *ServerInterfaceWrapper) GetSession(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetSession(c)
}

// GetPasteStatus operation middleware
func (siw *ServerInterfaceWrapper) GetPasteStatus(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetPasteStatus(c)
}

// CreatePaste operation middleware
func (siw *ServerInterfaceWrapper) CreatePaste(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreatePaste(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/read", wrapper.GetPaste)
	router.GET(options.BaseURL+"/session", wrapper.GetSession)
	router.POST(options.BaseURL+"/status", wrapper.GetPasteStatus)
	router.POST(options.BaseURL+"/submit", wrapper.CreatePaste)
}